+++ NEWS.md
@@ -1,5 +1,10 @@
 # rsconnect (development version)
 
+* `deployApp()` and `writeManifest()` now error if your library and lockfile
+  are out-of-sync. Previously it always used what was defined in the lockfile
+  but that was (a) slow and (b) could lead to different results than what you
+  see when running locally (#930).
+
 # rsconnect 1.0.1
 
 * `deployDoc()` includes `.Rprofile`, `requirements.txt` and `renv.lock` when