+++ NEWS.md
@@ -1,7 +1,7 @@
 # rsconnect (development version)
 
-* `deployApp()` and `writeManifest()` now error if your library and lockfile
-  are out-of-sync. Previously it always used what was defined in the lockfile
+* `deployApp()` and `writeManifest()` now error if your library and `renv.lock`
+  are out-of-sync. Previously it always used what was defined in the `renv.lock`
   but that was (a) slow and (b) could lead to different results than what you
   see when running locally (#930).