+++ NEWS.md
@@ -1,4 +1,4 @@
-# rsconnect (development version)
+# rsconnect 1.0.2
 
 * Fixed redeployments to shinyapps.io where `appName` is provided, but no local
   record of the deployment exists. (#932)